Add files via upload
I added Class objects for Cities and Goods, because there will be multiple of these, with multiple of their own attributes.
I added a variable for menuChoice; if the user's input for a menu choice is invalid, the terminal loop continues until the input is valid.
Next, I want to add a functional City Market. With every new day, there are different Goods with different prices, which can be bought and sold for the user's money.